Add Migration_Scheduler::do_background_migration()
As a convenience wrapper for both the 'do_background_migration' hook
and the is_migration_complete() method. Helpful as we need to move
the call to Migration_Scheduler::hook() out of schedule_migration()
as it's attached to 'shutdown', which occurs too late for it to be
attached and called as a callback, and we don't want to duplicate
the 'do_background_migration' hook.