Add Plugin::hook_scheduled_migration()
To make sure that Migration_Scheduler::hook() is called before the
action is triggered so that Migration_Scheduler::run() is called
when the action is triggered. As Plugin::schedule_migration() is
attached to 'shutdown', it was being called too late for the hook
to be attached correctly, meaning the background migration was never
run.